pyproject.toml
[tool.poetry]
name = "ProxyGPT"
version = "0.0.1"
description = ""
authors = ["Dante Noguez"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
redis = "^4.6.0"
fastapi = "^0.100.1"
python-dotenv = "^1.0.0"
uvicorn = "0.21.1"
gunicorn = "21.2.0"
openai = "^1.3.1"
httpx = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"